requirements.txt
discord.py==1.6.0
numpy==1.19.3
matplotlib==3.3.2
requests==2.24.0
emoji==1.2.0
asyncpg==0.21.0
parsedatetime==2.6
python-dateutil==2.8.1
aio-timers==0.0.1
Pillow==8.0.1